# specs/001-auth-keycloak-dashboard/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: StudyFlow API — Auth & Users
  version: 0.1.0
  description: |
    Contratos REST da feature 001-auth-keycloak-dashboard.
    Todas as rotas exigem Bearer JWT emitido pelo Keycloak (OIDC),
    exceto /health se exposto pelo runtime.
servers:
  - url: /api/v1
    description: API versionada
...
security:
  - bearerAuth: []
paths:
...
  /users/me:
    get:
      tags: [users]
      summary: Perfil do usuário autenticado
      responses:
        "200":
          description: Perfil
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Usuário desabilitado na base local
  /users/me/bootstrap:
    post:
      tags: [users]
      summary: Cria perfil local no primeiro acesso (auto-cadastro / sync)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserBootstrapRequest"
      responses:
        "201":
          description: Perfil criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "200":
          description: Perfil já existia — retorna atual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: E-mail ou CPF já cadastrado
  /users:
    get:
      tags: [users]
      summary: Lista usuários (Admin)
      parameters:
        - name: includeDisabled
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Lista
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [users]
      summary: Cria usuário no Keycloak + base (Admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateUserRequest"
      responses:
        "201":
          description: Criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Conflito de e-mail/CPF/Keycloak
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: JWT ausente ou inválido
    Forbidden:
      description: Perfil insuficiente ou usuário desabilitado
    BadRequest:
      description: Validação falhou
...
  schemas:
    ProfileType:
      type: string
      enum: [estudante, admin]
    EducationLevel:
      type: string
      enum: [fundamental, medio, graduacao, posGraduacao, outro]
    UserResponse:
      type: object
      required:
        - id
        - keycloakId
        - name
        - email
        - cpf
        - birthDate
        - educationLevel
        - profile
        - createdAt
        - updatedAt
        - disabled
      properties:
        id:
          type: string
          format: uuid
        keycloakId:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        cpf:
          type: string
          description: Pode ser mascarado em listagens (ex. ***.***.***-XX)
        birthDate:
          type: string
          format: date
        educationLevel:
          $ref: "#/components/schemas/EducationLevel"
        profile:
          $ref: "#/components/schemas/ProfileType"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        disabled:
          type: boolean
          description: true quando deletedAt está preenchido
    UserBootstrapRequest:
      type: object
      required: [name, cpf, birthDate, educationLevel]
      properties:
        name:
          type: string
        cpf:
          type: string
        birthDate:
          type: string
          format: date
        educationLevel:
          $ref: "#/components/schemas/EducationLevel"
    AdminCreateUserRequest:
      type: object
      required:
        - name
        - email
        - cpf
        - birthDate
        - educationLevel
        - profile
        - password
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        cpf:
          type: string
        birthDate:
          type: string
          format: date
        educationLevel:
          $ref: "#/components/schemas/EducationLevel"
        profile:
          $ref: "#/components/schemas/ProfileType"
        password:
          type: string
          format: password
          description: Senha final; nunca retornada; só enviada ao Keycloak
...